Using simple info instead of panel-item. Added wrap prop to simple-info
